Feeding

+++ wear/src/main/java/fr/croumy/bouge/presentation/data/dao/CompanionDao.kt
@@ -18,5 +18,5 @@ interface CompanionDao {
     fun insertCompanion(companionEntity: CompanionEntity)
 
     @Query("UPDATE companions SET happiness = :happiness, hungriness = :hungriness, health = :health WHERE deathDate IS NULL")
-    fun updateCompanionStats(happiness: Int, hungriness: Int, health: Float)
+    fun updateCompanionStats(happiness: Float, hungriness: Float, health: Float)
 }
