+++ wear/src/main/java/fr/croumy/bouge/presentation/models/companion/Stats.kt
@@ -1,7 +1,7 @@
 package fr.croumy.bouge.presentation.models.companion
 
 data class Stats(
-    val happiness: Int, // 0 TO 5
-    val hungriness: Int, // 0 TO 5
+    val happiness: Float, // 0 TO 5
+    val hungriness: Float, // 0 TO 5
     val health: Float, // 0 TO 5
 )
